refactor: 引入AnnotationTarget处理注解Target
