+++ spring-extension-commons/src/main/java/com/livk/commons/aop/AnnotationTarget.java
@@ -1,12 +1,9 @@
 /*
  * Copyright 2021-2024 spring-boot-extension the original author or authors.
- *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
  *       https://www.apache.org/licenses/LICENSE-2.0
- *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -14,23 +11,33 @@
  * limitations under the License.
  */
 
-package com.livk.local.lock;
+package com.livk.commons.aop;
 
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
+import com.google.common.collect.Sets;
+
+import java.lang.annotation.Annotation;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Target;
+import java.util.Set;
 
 /**
- * <p>
- * LocalLockExample
- * </p>
- *
  * @author livk
  */
-@SpringBootApplication
-public class LocalLockExample {
+final class AnnotationTarget<T extends Annotation> {
 
-	public static void main(String[] args) {
-		SpringApplication.run(LocalLockExample.class, args);
+	private final Set<ElementType> elementTypes;
+
+	public AnnotationTarget(Class<T> annotationType) {
+		Target target = annotationType.getAnnotation(Target.class);
+		this.elementTypes = Sets.newHashSet(target.value());
+	}
+
+	public boolean supportMethod() {
+		return elementTypes.contains(ElementType.METHOD);
+	}
+
+	public boolean supportType() {
+		return elementTypes.contains(ElementType.TYPE);
 	}
 
 }
